refactor(dts): split type value declaration emit
