+++ crates/tsz-emitter/src/declaration_emitter/core/mod.rs
@@ -5,6 +5,7 @@ mod emit_declaration_class_helpers;
 mod emit_declarations;
 mod emit_members;
 mod emit_statements;
+mod emit_type_value_declarations;
 mod import_rewrites;
 mod inferred_predicate_returns;
 mod js_class_static;
